refactor to default function
