+++ src/pages/AccessResultRecords.js
@@ -6,7 +6,7 @@ import { a, div, h, h3, h4, hr, i, label, span } from 'react-hyperscript-helpers
 import { ApplicationSummary } from '../components/ApplicationSummary';
 import { CollapsiblePanel } from '../components/CollapsiblePanel';
 import { CollectResultBox } from '../components/CollectResultBox';
-import { DataAccessRequest } from '../components/DataAccessRequest';
+import DataAccessRequestHeader from '../components/DataAccessRequestHeader';
 import { PageHeading } from '../components/PageHeading';
 import { SingleResultBox } from '../components/SingleResultBox';
 import TranslatedDULComponent from '../components/TranslatedDULComponent';
@@ -148,7 +148,7 @@ class AccessResultRecords extends Component {
               id: 'recordAccess', imgSrc: '/images/icon_access.png', iconSize: 'medium',
               color: 'access', title: 'Data Access - Results Record'
             }),
-            DataAccessRequest({
+            h(DataAccessRequestHeader, {
               isRendered: !ld.isEmpty(this.state.darInfo.datasets),
               dar: this.state.darInfo,
               consentName: this.state.consentName
