+++ src/pages/FinalAccessReview.js
@@ -7,7 +7,7 @@ import { ApplicationSummary } from '../components/ApplicationSummary';
 import { CollapsiblePanel } from '../components/CollapsiblePanel';
 import { CollectResultBox } from '../components/CollectResultBox';
 import { ConfirmationDialog } from '../components/ConfirmationDialog';
-import { DataAccessRequest } from '../components/DataAccessRequest';
+import DataAccessRequestHeader from '../components/DataAccessRequestHeader';
 import { PageHeading } from '../components/PageHeading';
 import { SingleResultBox } from '../components/SingleResultBox';
 import { SubmitVoteBox } from '../components/SubmitVoteBox';
@@ -574,7 +574,7 @@ class FinalAccessReview extends Component {
               id: 'finalAccess', imgSrc: '/images/icon_access.png', iconSize: 'medium',
               color: 'access', title: 'Final voting for Data Access Review'
             }),
-            DataAccessRequest({
+            h(DataAccessRequestHeader, {
               isRendered: !ld.isEmpty(this.state.darInfo.datasets),
               dar: this.state.darInfo,
               consentName: this.state.consentName
